feat: add tab for custom token dialog
