+++ packages/dashboard/src/pages/Wallets/components/AddTokenDialog/AddTokenFromList.tsx
@@ -1,0 +1,5 @@
+import { Box } from '@material-ui/core'
+
+export const AddTokenFromList = () => {
+    return <Box>token list</Box>
+}
